refactor(db): extract repository methods into db/repository.rs
