+++ src/db/mod.rs
@@ -8,8 +8,8 @@ use serde::{Deserialize, Serialize};
 use crate::migrations::{MIGRATIONS, Migration};
 use crate::model::{
     AgentStatus, Attachment, AttachmentOrigin, ChatMessage, ChatSession, CheckpointFile,
-    CompletedTurnData, ConversationCheckpoint, PinnedCommand, RemoteConnection, Repository,
-    TerminalTab, TurnToolActivity, Workspace, WorkspaceStatus,
+    CompletedTurnData, ConversationCheckpoint, PinnedCommand, RemoteConnection, TerminalTab,
+    TurnToolActivity, Workspace, WorkspaceStatus,
 };
 
 mod repository;
@@ -326,186 +326,6 @@ fn is_already_exists_error(err: &rusqlite::Error) -> bool {
 }
 
 impl Database {
-    // --- Repositories ---
-
-    pub fn insert_repository(&self, repo: &Repository) -> Result<(), rusqlite::Error> {
-        // New repos append at the end of the list.
-        let max_order: i32 = self
-            .conn
-            .query_row(
-                "SELECT COALESCE(MAX(sort_order), -1) FROM repositories",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap_or(-1);
-        self.conn.execute(
-            "INSERT INTO repositories (id, path, name, path_slug, sort_order)
-             VALUES (?1, ?2, ?3, ?4, ?5)",
-            params![repo.id, repo.path, repo.name, repo.path_slug, max_order + 1],
-        )?;
-        Ok(())
-    }
-
-    fn parse_repo_row(row: &rusqlite::Row) -> rusqlite::Result<Repository> {
-        Ok(Repository {
-            id: row.get(0)?,
-            path: row.get(1)?,
-            name: row.get(2)?,
-            icon: row.get(3)?,
-            path_slug: row.get::<_, Option<String>>(4)?.unwrap_or_default(),
-            created_at: row.get(5)?,
-            setup_script: row.get(6)?,
-            custom_instructions: row.get(7)?,
-            sort_order: row.get(8)?,
-            branch_rename_preferences: row.get(9)?,
-            setup_script_auto_run: row.get::<_, i32>(10).unwrap_or(0) != 0,
-            base_branch: row.get(11)?,
-            default_remote: row.get(12)?,
-            path_valid: true, // validated after load
-        })
-    }
-
-    pub fn list_repositories(&self) -> Result<Vec<Repository>, rusqlite::Error> {
-        let mut stmt = self.conn.prepare(
-            "SELECT id, path, name, icon, path_slug, created_at, setup_script, custom_instructions, sort_order, branch_rename_preferences, setup_script_auto_run, base_branch, default_remote
-             FROM repositories ORDER BY sort_order, name",
-        )?;
-        let rows = stmt.query_map([], Self::parse_repo_row)?;
-        rows.collect()
-    }
-
-    pub fn get_repository(&self, id: &str) -> Result<Option<Repository>, rusqlite::Error> {
-        self.conn
-            .query_row(
-                "SELECT id, path, name, icon, path_slug, created_at, setup_script, custom_instructions, sort_order, branch_rename_preferences, setup_script_auto_run, base_branch, default_remote
-                 FROM repositories WHERE id = ?1",
-                params![id],
-                Self::parse_repo_row,
-            )
-            .optional()
-    }
-
-    pub fn update_repository_path(&self, id: &str, path: &str) -> Result<(), rusqlite::Error> {
-        self.conn.execute(
-            "UPDATE repositories SET path = ?1 WHERE id = ?2",
-            params![path, id],
-        )?;
-        Ok(())
-    }
-
-    /// Batch-update sort_order for repositories based on the provided ID order.
-    pub fn reorder_repositories(&self, ids: &[String]) -> Result<(), rusqlite::Error> {
-        let tx = self.conn.unchecked_transaction()?;
-        {
-            let mut stmt = tx.prepare("UPDATE repositories SET sort_order = ?1 WHERE id = ?2")?;
-            for (i, id) in ids.iter().enumerate() {
-                stmt.execute(params![i as i32, id])?;
-            }
-        }
-        tx.commit()?;
-        Ok(())
-    }
-
-    pub fn delete_repository(&self, id: &str) -> Result<(), rusqlite::Error> {
-        self.conn
-            .execute("DELETE FROM repositories WHERE id = ?1", params![id])?;
-        Ok(())
-    }
-
-    #[allow(dead_code)]
-    pub fn update_repository_name(&self, id: &str, name: &str) -> Result<(), rusqlite::Error> {
-        self.conn.execute(
-            "UPDATE repositories SET name = ?1 WHERE id = ?2",
-            params![name, id],
-        )?;
-        Ok(())
-    }
-
-    #[allow(dead_code)]
-    pub fn update_repository_icon(
-        &self,
-        id: &str,
-        icon: Option<&str>,
-    ) -> Result<(), rusqlite::Error> {
-        self.conn.execute(
-            "UPDATE repositories SET icon = ?1 WHERE id = ?2",
-            params![icon, id],
-        )?;
-        Ok(())
-    }
-
-    pub fn update_repository_setup_script(
-        &self,
-        id: &str,
-        script: Option<&str>,
-    ) -> Result<(), rusqlite::Error> {
-        self.conn.execute(
-            "UPDATE repositories SET setup_script = ?1 WHERE id = ?2",
-            params![script, id],
-        )?;
-        Ok(())
-    }
-
-    pub fn update_repository_setup_script_auto_run(
-        &self,
-        id: &str,
-        enabled: bool,
-    ) -> Result<(), rusqlite::Error> {
-        self.conn.execute(
-            "UPDATE repositories SET setup_script_auto_run = ?1 WHERE id = ?2",
-            params![enabled as i32, id],
-        )?;
-        Ok(())
-    }
-
-    pub fn update_repository_base_branch(
-        &self,
-        id: &str,
-        base_branch: Option<&str>,
-    ) -> Result<(), rusqlite::Error> {
-        self.conn.execute(
-            "UPDATE repositories SET base_branch = ?1 WHERE id = ?2",
-            params![base_branch, id],
-        )?;
-        Ok(())
-    }
-
-    pub fn update_repository_default_remote(
-        &self,
-        id: &str,
-        default_remote: Option<&str>,
-    ) -> Result<(), rusqlite::Error> {
-        self.conn.execute(
-            "UPDATE repositories SET default_remote = ?1 WHERE id = ?2",
-            params![default_remote, id],
-        )?;
-        Ok(())
-    }
-
-    pub fn update_repository_custom_instructions(
-        &self,
-        id: &str,
-        instructions: Option<&str>,
-    ) -> Result<(), rusqlite::Error> {
-        self.conn.execute(
-            "UPDATE repositories SET custom_instructions = ?1 WHERE id = ?2",
-            params![instructions, id],
-        )?;
-        Ok(())
-    }
-
-    pub fn update_repository_branch_rename_preferences(
-        &self,
-        id: &str,
-        preferences: Option<&str>,
-    ) -> Result<(), rusqlite::Error> {
-        self.conn.execute(
-            "UPDATE repositories SET branch_rename_preferences = ?1 WHERE id = ?2",
-            params![preferences, id],
-        )?;
-        Ok(())
-    }
-
     // --- App Settings ---
 
     #[allow(dead_code)]
@@ -2486,7 +2306,7 @@ impl Database {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::model::{AgentStatus, Attachment, ChatRole, WorkspaceStatus};
+    use crate::model::{AgentStatus, Attachment, ChatRole, Repository, WorkspaceStatus};
 
     fn make_repo(id: &str, path: &str, name: &str) -> Repository {
         Repository {
